Use existing cypress backend
