--- e2e/embedding-sdk-host-apps/vite-5-host-app/.env.docker.example
@@ -1,7 +1,0 @@
-PREMIUM_EMBEDDING_TOKEN=""
-METABASE_EMBEDDING_API_KEY="mb_abcdefghijklmnopqrstuvwxyz0123456789"
-
-MB_PORT=4300
-CLIENT_PORT=4400
-
-WATCH="false"
