--- e2e/embedding-sdk-host-apps/vite-5-host-app/e2e/support/cypress.config.js
@@ -1,12 +1,0 @@
-const path = require("path");
-
-const { defineConfig } = require("cypress");
-
-module.exports = defineConfig({
-  e2e: {
-    baseUrl: `http://localhost:${process.env.CLIENT_PORT}`,
-    supportFile: path.resolve(path.join(__dirname, "./cypress.js")),
-    specPattern: path.resolve(path.join(__dirname, "../test/**/*.cy.spec.js")),
-    defaultBrowser: "chrome",
-  },
-});
