--- e2e/embedding-sdk-host-apps/vite-5-host-app/e2e/support/cypress.js
@@ -1,1 +1,0 @@
-import "@testing-library/cypress/add-commands";
