--- e2e/embedding-sdk-host-apps/vite-5-host-app/metabase/Dockerfile
@@ -1,11 +1,0 @@
-FROM metabase/metabase-enterprise:v1.54.x
-
-COPY ./metabase /app/
-COPY ./local-dist /app/local-dist
-
-RUN if [ -f ./app/local-dist/metabase.jar ]; then \
-      echo "Local metabase.jar is found in ./app/local-dist, running it..."; \
-      cp ./app/local-dist/metabase.jar /app/metabase.jar; \
-    else \
-      echo "Local metabase.jar is not found in ./app/local-dist, skipping copy"; \
-    fi
