--- e2e/embedding-sdk-host-apps/vite-5-host-app/metabase/init-config.yml
@@ -1,15 +1,0 @@
-version: 1
-config:
-  users:
-    - first_name: Rene
-      last_name: Descartes
-      password: foobarbaz
-      email: rene@example.com
-  api-keys:
-    - name: "Admin API key"
-      group: admin
-      creator: rene@example.com
-      key: "{{ env METABASE_EMBEDDING_API_KEY }}"
-  settings:
-    enable-embedding-sdk: true
-    jwt-enabled: true
