+++ e2e/runner/constants/backend-port.js
@@ -1,0 +1,1 @@
+export const BACKEND_PORT = process.env.BACKEND_PORT ?? 4000;
