+++ e2e/support/helpers/embedding-sdk-component-testing/component-embedding-sdk-question-helpers.tsx
