+++ e2e/support/helpers/embedding-sdk-component-testing/index.ts
