+++ e2e/support/helpers/embedding-sdk-helpers/constants.ts
@@ -1,0 +1,4 @@
+export const METABASE_INSTANCE_URL = "http://localhost:4000";
+export const AUTH_PROVIDER_URL = "http://auth-provider/sso";
+export const JWT_SHARED_SECRET =
+  "0000000000000000000000000000000000000000000000000000000000000000";
