+++ e2e/support/helpers/embedding-sdk-testing/index.ts
@@ -1,0 +1,1 @@
+export * from ".//embedding-sdk-helpers";
