+++ e2e/test-component/scenarios/embedding-sdk/interactive-question-map.cy.spec.tsx
@@ -1,12 +1,13 @@
 import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import { createQuestion, mapPinIcon } from "e2e/support/helpers";
 import {
   METABASE_INSTANCE_URL,
-  mockAuthProviderAndJwtSignIn,
-  mountInteractiveQuestion,
-  signInAsAdminAndEnableEmbeddingSdk,
-} from "e2e/support/helpers/component-testing-sdk";
+  createQuestion,
+  mapPinIcon,
+} from "e2e/support/helpers";
 import { getSdkRoot } from "e2e/support/helpers/e2e-embedding-sdk-helpers";
+import { mountInteractiveQuestion } from "e2e/support/helpers/embedding-sdk-component-testing";
+import { signInAsAdminAndEnableEmbeddingSdk } from "e2e/support/helpers/embedding-sdk-testing";
+import { mockAuthProviderAndJwtSignIn } from "e2e/support/helpers/embedding-sdk-testing/embedding-sdk-helpers";
 
 const { PEOPLE_ID } = SAMPLE_DATABASE;
