+++ e2e/test-component/scenarios/embedding-sdk/interactive-question-native.cy.spec.tsx
@@ -3,11 +3,9 @@ import {
   createNativeQuestion,
   tableInteractiveBody,
 } from "e2e/support/helpers";
-import {
-  mockAuthProviderAndJwtSignIn,
-  mountInteractiveQuestion,
-  signInAsAdminAndEnableEmbeddingSdk,
-} from "e2e/support/helpers/component-testing-sdk";
+import { mountInteractiveQuestion } from "e2e/support/helpers/embedding-sdk-component-testing";
+import { signInAsAdminAndEnableEmbeddingSdk } from "e2e/support/helpers/embedding-sdk-testing";
+import { mockAuthProviderAndJwtSignIn } from "e2e/support/helpers/embedding-sdk-testing/embedding-sdk-helpers";
 import type { DatasetColumn } from "metabase-types/api";
 
 const { ORDERS, ORDERS_ID } = SAMPLE_DATABASE;
@@ -40,7 +38,7 @@ describe("scenarios > embedding-sdk > interactive-question > native", () => {
     mockAuthProviderAndJwtSignIn();
   });
 
-  it("supports passing sql parameters to native questions", () => {
+  it("e2e/supports passing sql parameters to native questions", () => {
     mountInteractiveQuestion({ initialSqlParameters: { ID: ORDERS_ID } });
 
     cy.wait("@cardQuery").then(({ response }) => {
