+++ e2e/test-component/scenarios/embedding-sdk/locale-on-provider.cy.spec.tsx
@@ -4,13 +4,17 @@ import {
 } from "@metabase/embedding-sdk-react";
 
 import { ORDERS_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
+import { METABASE_INSTANCE_URL } from "e2e/support/helpers";
 import { updateSetting } from "e2e/support/helpers/api";
 import {
+  AUTH_PROVIDER_URL,
   METABASE_INSTANCE_URL,
   mockAuthProviderAndJwtSignIn,
   signInAsAdminAndEnableEmbeddingSdk,
 } from "e2e/support/helpers/component-testing-sdk";
 import { getSdkRoot } from "e2e/support/helpers/e2e-embedding-sdk-helpers";
+import { signInAsAdminAndEnableEmbeddingSdk } from "e2e/support/helpers/embedding-sdk-testing";
+import { mockAuthProviderAndJwtSignIn } from "e2e/support/helpers/embedding-sdk-testing/embedding-sdk-helpers";
 
 function setup({
   locale,
