+++ package.json
@@ -252,7 +252,6 @@
     "cypress-real-events": "^1.14.0",
     "cypress-split": "^1.24.14",
     "cypress-terminal-report": "^6.1.2",
-    "dotenv-cli": "^8.0.0",
     "esbuild": "^0.25.0",
     "eslint": "7.32.0",
     "eslint-import-resolver-webpack": "^0.13.10",
